IN-348 fix: fix infinity scroll with rtk forceRefetch
